backend/requirements.txt
# Launch Transportation Management Platform - Backend Dependencies
# Copyright (c) 2025 Robert Vassallo. All rights reserved.
# This software is proprietary and confidential.
#
# Django Backend Requirements for Launch TMS
# Core Django
Django==5.0.4
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1

# API Documentation
drf-spectacular==0.27.2

# File handling
Pillow==10.3.0
django-storages==1.14.2

# Development
python-decouple==3.8
django-extensions==3.2.3

# Testing
pytest-django==4.8.0
factory-boy==3.3.0

# Utilities
python-dateutil==2.9.0
requests==2.31.0

# Fix for pkg_resources deprecation warning
setuptools<81
